refactor(storage-sqlite): extract noteMapping helpers from SQLiteNoteRepository
Phase 1 of the SQLiteNoteRepository split. Pure helpers come out to
their own file so future sub-repos (sync, tag, archive) can reuse them
without inheriting from the 1000-line class.

SQLiteNoteRepository.ts: 1121 -> 1038 lines (-7%).
noteMapping.ts: new (133 lines).

Extracted:
- NoteRow / TagRow / TagWithColorRow / BacklinkInfo  — row types
- rowToNote(row, tags) -> Note                       — pure mapper
- prepareFtsQuery(input) -> string                   — FTS5 query escaper
- archivedConditionSql(filter, alias) -> string      — SQL fragment helper

Call sites swapped from `this.<helper>()` to plain function imports.
The public class signature is unchanged; `BacklinkInfo` is re-exported
so external consumers (apps/desktop/handlers/types.ts) keep working.

What this PR DELIBERATELY does NOT do:

- Extract sync methods (getPendingChanges through getSyncHistory ~430
  lines) into a SQLiteNoteSyncRepository. Those share state (tag
  queries, transactions, FTS sync triggers) with the main class in
  ways that need real-DB integration coverage to refactor safely. The
  helpers extracted here are the foundation: a follow-up PR can build
  the sync sub-repo on top of them without touching the helpers again.
- Extract tag methods (setManualTags, renameTag, getAllTagsWithColors,
  etc.) for the same reason.

Validates:
- pnpm -r typecheck — green
- pnpm test — 17/17 packages (storage-sqlite tests skipped per repo
  policy; helpers are pure and trivially exercised by the desktop's
  integration tests)
